cmd: make auth token optional; default to store
